requirements.txt
# Core dependencies
langextract>=1.0.0
click>=8.0.0
python-dotenv>=0.19.0
PyYAML>=6.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pymupdf>=1.23.0

# Data processing and output
pandas>=1.5.0
openpyxl>=3.1.0
json5>=0.9.0

# Utilities
rich>=13.0.0
tqdm>=4.64.0
pathlib2>=2.3.0

# Optional: For advanced PDF processing
# tesseract>=0.1.3  # OCR support
# Pillow>=9.0.0     # Image processing
